Make 64-bit to int narrowing explicit outside macros
Clang reported 602 implicit conversions from 64-bit types (size_t, long,
bitvector_t, uint64_t, time_t, MySQL row counts) to int, and GCC 16.2
reported the same sites under -Wconversion. asciiflag_conv results are
stored one 32-bit flag element at a time, so those assignments now cast
to int. The other sites are cast at the expression Clang identifies as
converted, only where that expression is outside every macro, so each
cast performs exactly the conversion that already happened. Experience,
gold and clan treasure are capped below INT_MAX, and the object
editor's max_val, whose largest value is 400000000, is now int.

GCC 16.2 conversion falls from 1258 to 863 sites and Clang 22.1.8
shorten-64-to-32 from 602 to 205.

+++ src/account.c
@@ -1329,8 +1329,8 @@ void show_account_menu(struct descriptor_data *d)
                   if (CLASS_LEVEL(tch, inc))
                   {
                     if (classCount)
-                      len = snprintf_append(buf, sizeof(buf), len, "/");
-                    len = snprintf_append(buf, sizeof(buf), len, "%s", CLSLIST_CLRABBRV(inc));
+                      len = snprintf_append(buf, sizeof(buf), (int)len, "/");
+                    len = snprintf_append(buf, sizeof(buf), (int)len, "%s", CLSLIST_CLRABBRV(inc));
                     classCount++;
                   }
                 }
